Implemented lot more methods;

+++ src/entities/User.hpp
@@ -5,7 +5,7 @@
 namespace CppTelegramBots {
     class User {
     public:
-        long long id{};
+        int id{};
 
         std::string firstName{};
         std::string lastName{};
